Updated Job, UtilDB, and created UpdateJob
Job was missing a check for if the jobs were open, closed, or deleted,
so that is added and all related returns are added as well. Added a
filter for personal jobs and a sort by value for Bids. UpdateJobs is
done. Need to make it so the customer only has two options to update on
the front end.